feat: adds render command and changes build command to store collections in cache
The command structure has been changed to resemble podman/docker build
workflows of build and push.

Closes #32
