+++ cli/cli_e2e_test.go
@@ -44,9 +44,10 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
-				RootDir: "testdata/flatworkspace",
+				RootDir:     "testdata/flatworkspace",
+				Destination: fmt.Sprintf("%s/client-flat-test:latest", u.Host),
 			},
 			pushOpts: &PushOptions{
 				RootOptions: &RootOptions{
@@ -55,7 +56,7 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
 				Destination: fmt.Sprintf("%s/client-flat-test:latest", u.Host),
 				PlainHTTP:   true,
@@ -67,7 +68,7 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
 				PlainHTTP: true,
 				Output:    t.TempDir(),
@@ -82,9 +83,10 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
-				RootDir: "testdata/multi-level-workspace",
+				RootDir:     "testdata/multi-level-workspace",
+				Destination: fmt.Sprintf("%s/client-multi-test:latest", u.Host),
 			},
 			pushOpts: &PushOptions{
 				RootOptions: &RootOptions{
@@ -93,7 +95,7 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
 				Destination: fmt.Sprintf("%s/client-multi-test:latest", u.Host),
 				PlainHTTP:   true,
@@ -105,7 +107,7 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
 				Output:    t.TempDir(),
 				PlainHTTP: true,
@@ -120,9 +122,10 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
-				RootDir: "testdata/uor-template",
+				RootDir:     "testdata/uor-template",
+				Destination: fmt.Sprintf("%s/client-uor-test:latest", u.Host),
 			},
 			pushOpts: &PushOptions{
 				RootOptions: &RootOptions{
@@ -131,7 +134,7 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
 				Destination: fmt.Sprintf("%s/client-uor-test:latest", u.Host),
 				PlainHTTP:   true,
@@ -143,7 +146,7 @@ func TestCLIE2E(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger:   testlogr,
+					Logger: testlogr,
 				},
 				Output:    t.TempDir(),
 				PlainHTTP: true,
@@ -153,7 +156,13 @@ func TestCLIE2E(t *testing.T) {
 
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
-			c.buildOpts.Output = t.TempDir()
+
+			cache := filepath.Join(t.TempDir(), "cache")
+			require.NoError(t, os.MkdirAll(cache, 0750))
+			c.pushOpts.cacheDir = cache
+			c.pullOpts.cacheDir = cache
+			c.buildOpts.cacheDir = cache
+
 			err := c.buildOpts.Run(context.TODO())
 			if c.expBuildError != "" {
 				require.EqualError(t, err, c.expBuildError)
@@ -161,12 +170,6 @@ func TestCLIE2E(t *testing.T) {
 				require.NoError(t, err)
 			}
 
-			cache := filepath.Join(t.TempDir(), "cache")
-			require.NoError(t, os.MkdirAll(cache, 0750))
-			c.pushOpts.cacheDir = cache
-			c.pullOpts.cacheDir = cache
-
-			c.pushOpts.RootDir = c.buildOpts.Output
 			err = c.pushOpts.Run(context.TODO())
 			if c.expPushError != "" {
 				require.EqualError(t, err, c.expBuildError)
