+++ cli/push_test.go
@@ -2,15 +2,10 @@ package cli
 
 import (
 	"context"
-	"fmt"
 	"io/ioutil"
-	"net/http/httptest"
-	"net/url"
 	"os"
-	"path/filepath"
 	"testing"
 
-	"github.com/google/go-containerregistry/pkg/registry"
 	"github.com/stretchr/testify/require"
 	"github.com/uor-framework/uor-client-go/cli/log"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
@@ -28,9 +23,8 @@ func TestPushComplete(t *testing.T) {
 	cases := []spec{
 		{
 			name: "Valid/CorrectNumberOfArguments",
-			args: []string{"testdata", "test-registry.com/image:latest"},
+			args: []string{"test-registry.com/image:latest"},
 			expOpts: &PushOptions{
-				RootDir:     "testdata",
 				Destination: "test-registry.com/image:latest",
 			},
 			opts: &PushOptions{},
@@ -40,7 +34,7 @@ func TestPushComplete(t *testing.T) {
 			args:     []string{},
 			expOpts:  &PushOptions{},
 			opts:     &PushOptions{},
-			expError: "bug: expecting two arguments",
+			expError: "bug: expecting one argument",
 		},
 	}
 
@@ -57,50 +51,10 @@ func TestPushComplete(t *testing.T) {
 	}
 }
 
-func TestPushValidate(t *testing.T) {
-	type spec struct {
-		name     string
-		opts     *PushOptions
-		expError string
-	}
-
-	cases := []spec{
-		{
-			name: "Valid/RootDirExists",
-			opts: &PushOptions{
-				RootDir: "testdata",
-			},
-		},
-		{
-			name: "Invalid/RootDirDoesNotExist",
-			opts: &PushOptions{
-				RootDir: "fake",
-			},
-			expError: "workspace directory \"fake\": stat fake: no such file or directory",
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			err := c.opts.Validate()
-			if c.expError != "" {
-				require.EqualError(t, err, c.expError)
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
 func TestPushRun(t *testing.T) {
 	testlogr, err := log.NewLogger(ioutil.Discard, "debug")
 	require.NoError(t, err)
 
-	server := httptest.NewServer(registry.New())
-	t.Cleanup(server.Close)
-	u, err := url.Parse(server.URL)
-	require.NoError(t, err)
-
 	type spec struct {
 		name     string
 		opts     *PushOptions
@@ -109,7 +63,7 @@ func TestPushRun(t *testing.T) {
 
 	cases := []spec{
 		{
-			name: "Success/FlatWorkspace",
+			name: "Failure/NotStored",
 			opts: &PushOptions{
 				RootOptions: &RootOptions{
 					IOStreams: genericclioptions.IOStreams{
@@ -117,52 +71,19 @@ func TestPushRun(t *testing.T) {
 						In:     os.Stdin,
 						ErrOut: os.Stderr,
 					},
-					Logger: testlogr,
+					Logger:   testlogr,
+					cacheDir: "testdata/cache",
 				},
-				Destination: fmt.Sprintf("%s/client-flat-test:latest", u.Host),
-				RootDir:     "testdata/flatworkspace",
+				Destination: "locahost:5001/client-flat-test:latest",
 				PlainHTTP:   true,
 			},
-		},
-		{
-			name: "Success/MultiLevelWorkspace",
-			opts: &PushOptions{
-				RootOptions: &RootOptions{
-					IOStreams: genericclioptions.IOStreams{
-						Out:    os.Stdout,
-						In:     os.Stdin,
-						ErrOut: os.Stderr,
-					},
-					Logger: testlogr,
-				},
-				Destination: fmt.Sprintf("%s/client-multi-test:latest", u.Host),
-				RootDir:     "testdata/multi-level-workspace",
-				PlainHTTP:   true,
-			},
-		},
-		{
-			name: "SuccessTwoRoots",
-			opts: &PushOptions{
-				RootOptions: &RootOptions{
-					IOStreams: genericclioptions.IOStreams{
-						Out:    os.Stdout,
-						In:     os.Stdin,
-						ErrOut: os.Stderr,
-					},
-					Logger: testlogr,
-				},
-				Destination: fmt.Sprintf("%s/client-tworoots-test:latest", u.Host),
-				RootDir:     "testdata/tworoots",
-				PlainHTTP:   true,
-			},
+			expError: "error publishing content to locahost:5001/client-flat-test:latest:" +
+				" descriptor for reference locahost:5001/client-flat-test:latest is not stored",
 		},
 	}
 
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
-			cache := filepath.Join(t.TempDir(), "cache")
-			require.NoError(t, os.MkdirAll(cache, 0750))
-			c.opts.cacheDir = cache
 			err := c.opts.Run(context.TODO())
 			if c.expError != "" {
 				require.EqualError(t, err, c.expError)
