+++ cli/root.go
@@ -66,7 +66,8 @@ func NewRootCmd() *cobra.Command {
 		"Log level (debug, info, warn, error, fatal)")
 
 	// TODO(sabre1041) Reenable/remove once build capability strategy determined
-	//cmd.AddCommand(NewBuildCmd(&o))
+	//cmd.AddCommand(NewRenderCmd(&o))
+	cmd.AddCommand(NewBuildCmd(&o))
 	cmd.AddCommand(NewPushCmd(&o))
 	cmd.AddCommand(NewPullCmd(&o))
 	cmd.AddCommand(NewVersionCmd(&o))
