+++ cli/testdata/cache/oci-layout
@@ -1,0 +1,1 @@
+{"imageLayoutVersion":"1.0.0"}
