+++ content/storage.go
@@ -6,7 +6,7 @@ import (
 
 // Store defines the methods for add, inspecting, and removing
 // OCI content for a storage location. The interface wraps oras
-// Storage and TagResolver interfaces for use with `oras` copy methods.
+// Storage and TagResolver interfaces for use with `oras` Copy methods.
 type Store interface {
 	// Storage represents a content-addressable storage where contents are
 	// accessed via Descriptors.
