+++ registryclient/client.go
@@ -4,11 +4,33 @@ import (
 	"context"
 
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+
+	"github.com/uor-framework/uor-client-go/content"
 )
 
 // Client defines methods to interact with OCI artifacts
 // in various contexts.
 type Client interface {
+	Remote
+	Local
+}
+
+// Remote defines methods to interact with OCI
+// artifacts in remote contexts.
+type Remote interface {
+	// Push pushes an artifact to a remote registry from a source
+	// content store.
+	Push(context.Context, content.Store, string) (ocispec.Descriptor, error)
+	// Pull pulls an artifact from a remote registry to a local
+	// content store.
+	Pull(context.Context, string, content.Store) (ocispec.Descriptor, error)
+}
+
+// Local defines methods to interact with OCI artifacts
+// in a local context. An underlying store can be used to store
+// each descriptor and is return in the Store method for use with
+// Push and Pull operations or oras Copy operations.
+type Local interface {
 	// GatherDescriptors loads files to create OCI descriptors with a specific
 	// media type.
 	GatherDescriptors(context.Context, string, ...string) ([]ocispec.Descriptor, error)
@@ -18,25 +40,12 @@ type Client interface {
 	// GenerateManifest creates and stores a manifest for an image reference.
 	// This is generated from the config descriptor and artifact descriptors.
 	GenerateManifest(context.Context, string, ocispec.Descriptor, map[string]string, ...ocispec.Descriptor) (ocispec.Descriptor, error)
-	// Execute performs the copy of OCI artifacts.
-	// The image reference and copy action are expected as inputs.
-	Execute(context.Context, string, ActionType) (ocispec.Descriptor, error)
-	// Destroy cleans up any on-disk resources used to track descriptors.
+	// Save saves a built artifact to local store.
+	Save(context.Context, string, content.Store) (ocispec.Descriptor, error)
+	// Store returns the underlying content store
+	// used for OCI artifact building.
+	Store() (content.Store, error)
+	// Destroy cleans up temporary files on-disk
+	// for tracking descriptors
 	Destroy() error
 }
-
-// ActionType defines what actions (e.g. push, push, etc...) the execute method should perform.
-type ActionType int
-
-const (
-	// TypeInvalid is the default action.
-	// It is invalid because the action must
-	// be explicitly set.
-	TypeInvalid ActionType = iota
-	// TypePush action pushes from a local location
-	// to a remote location.
-	TypePush
-	// TypePull action pulls from a remote location to
-	// a local location.
-	TypePull
-)
